changed var type

+++ 0x19-hash_tables/2-key_index.c
@@ -10,7 +10,7 @@
 unsigned long int key_index(const unsigned char *key, unsigned long int size)
 {
 	unsigned long int hash;
-	int c;
+	unsigned long int c;
 
 	hash = hash_djb2(key);
 	c = hash % size;
